Updated endpoint to rss.xml
